Refactor TaskScreenContainer to use hoisted state and move to tasks package
